Move Entities and Repository classes into appropriate folders
