--- src/MyProject/Bundle/MainBundle/Entity/User.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace MyProject\Bundle\MainBundle\Entity;
-
-use FOS\UserBundle\Model\User as BaseUser;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- */
-class User extends BaseUser
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-}
