+++ src/Repository/UserRepository.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace RCatlin\Blog\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+class UserRepository extends EntityRepository
+{
+}
